// src/app/openrqm-api/models/RQMDocument.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RQMDocument = {
    id?: number;
    workspaceId?: number;
    internalIdentifier?: number;
    externalIdentifier?: string;
    name?: string;
    shortName?: string;
    description?: string;
    confidentiality?: string;
    authorId?: number;
    languageId?: number;
    approverId?: number;
    reviewerText?: string;
    lastModifiedById?: number;
    lastModifiedOn?: string;
    baselineMajor?: number;
    baselineMinor?: number;
    baselineReview?: number;
    previousBaselineId?: number;
};